Move logic from App.js to Board component
