+++ src/components/Tetromino/Tetromino.css
@@ -1,3 +1,3 @@
-.block__wrapper {
+.Tetromino {
     position: relative;
 }
